refactor: layout is refactored into new file.
